+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# UNRELEASED
+
+ - Allow passing arbitrary data to list(), retrieve(), and delete()
+
+
 # 0.1.0 (2017/03/27)
 
  - Add list() method